Add files via upload
The Mental Health dataset is a collection of texts about people dealing with anxiety, depression, and other mental health disorders. The dataset is divided into two columns: one with the comments and another with labels indicating whether the comments are poisonous or not.It can be utilised for a range of tasks, including sentiment analysis, toxic language identification, and mental health language analysis. The  data may be beneficial to scholars, mental health providers, and others who want to study the language and attitude surrounding mental health issues.